Move kv_manager into a directory and decouple tests
